requirements.txt
Flask==3.0.0
numpy==1.26.2
matplotlib==3.8.2
gunicorn==21.2.0
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Migrate==4.0.5
